backend/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
gunicorn==20.0.4
asgiref==3.4.1
backcall==0.2.0
dataclasses
decorator==5.0.9
Django==3.0.5
djangorestframework==3.13.1
djongo==1.3.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
pymongo==3.12.0
pytz==2021.1
six==1.16.0
sqlparse==0.2.4
traitlets==4.3.3
typing-extensions==3.10.0.2
wcwidth==0.2.5
whitenoise
bcrypt==3.2.2
requests
bson
pyjwt
django-rest-swagger
drf_yasg